Move some struct upward in cuda_runtime_api.cc
